misc: offload `buildClassAttrib` into ids-string-utils for re-use between ids-tab & ids-tabs
