+++ src/ids-base/index.js
@@ -1,3 +1,4 @@
 export { default as IdsThemeMixin } from './ids-theme-mixin';
 export { default as IdsEventsMixin } from './ids-events-mixin';
 export { default as IdsKeyboardMixin } from './ids-keyboard-mixin';
+export { default as IdsStringUtils } from './ids-string-utils';
